Put and patch routes and db operations complete for users

+++ backend/app/database/repository.py
@@ -1,6 +1,6 @@
 from typing import Generic, TypeVar
 
-from sqlalchemy import BinaryExpression, select
+from sqlalchemy import BinaryExpression, select, update
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.models.base import Base
@@ -22,15 +22,22 @@ class DBRepository(Generic[Model]):
     async def get(self, id: int) -> Model | None:
         return await self.session.get(self.model, id)
     
-    # async def update(self, data: dict) -> Model:
-    #     instance = self.model(**data)
-    #     self.session.add(instance)
-    #     await self.session.commit()
-    #     await self.session.refresh(instance)
-    #     return instance
-    
-    async def delete(self, id: int) -> str:
+    async def update(self, id: int, data: dict) -> Model | None:
         instance = await self.session.get(self.model, id)
+        if instance is None:
+            return None
+        for key, value in data.items():
+            setattr(instance, key, value)
+        self.session.add(instance)
+        await self.session.commit()
+        await self.session.refresh(instance)
+        return instance
+    
+    
+    async def delete(self, id: int) -> str | None:
+        instance = await self.session.get(self.model, id)
+        if instance is None:
+            return None
         await self.session.delete(instance)
         await self.session.commit()
         return f'User: {id} deleted'
